requirements.txt
python-dotenv
schedule
smtplib

